SA/oai_subjective_analysis/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.96.0
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.1.0
h11==0.14.0
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector-python==8.0.33
nltk==3.8.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai==0.27.8
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
pydantic==1.10.8
python-multipart==0.0.6
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.6.3
urllib3==1.26.15
uvicorn==0.22.0
yarl==1.9.2
zipp==3.15.0
